Changed naming convention of interfaces
