+++ src/model/game_pieces/i_dev_card.py
@@ -1,0 +1,13 @@
+from abc import ABC, abstractmethod
+from model.item.item import Item
+from model.encounter.encounter import Encounter
+
+class IDevCard(ABC):
+
+    @abstractmethod
+    def get_item(self) -> Item:
+        pass
+
+    @abstractmethod
+    def get_encounter(self, time: int) -> Encounter:
+        pass
