--- src/model/game_pieces/mock_interfaces.py
@@ -1,5 +1,0 @@
-class Item:
-    pass
-
-class Event:
-    pass
